some tests

+++ backend/src/main/java/recipes/fridger/backend/controller/UserController.java
@@ -238,10 +238,10 @@ public class UserController {
     public @ResponseBody Iterable<Goal>
     getGoals(@RequestParam(required = false) Long id,
                 @RequestParam(required = false) String endGoal,
-                @RequestParam(required = false) Double calories,
-                @RequestParam(required = false) Double carbs,
-                @RequestParam(required = false) Double protein,
-                @RequestParam(required = false) Double fat,
+                @RequestParam(required = false) Integer calories,
+                @RequestParam(required = false) Integer carbs,
+                @RequestParam(required = false) Integer protein,
+                @RequestParam(required = false) Integer fat,
                 @RequestParam(required = false) Double currWeight,
                 @RequestParam(required = false) Double goalWeight,
                 @RequestParam(required = false) Long userId)
